favorite location features
implemented the favorite location feature to the program. user must be logged in and have a favorite zip code identified. if these conditions are met, options 1, 6, 10 and 19 will automatically return results with that favorite zip code. under any other conditions, the user will be prompted to enter a zip code.